コードレビューの指摘に従い修正
